Add MazeShortestPath
Given a maze in the form of the binary rectangular matrix, find the length of the shortest path in maze from a given source to a given destination. The path can only be constructed out of cells having value 1 and at any given moment, we can only move one step in one of the four directions.